refactor: Clarify role of internals module
