+++ src/branch/mod.rs
@@ -5,7 +5,7 @@ mod tests;
 
 use crate::error::ErrorKind;
 use crate::error::ParseError;
-use crate::internal::{Err, IResult, Parser};
+use crate::{Err, IResult, Parser};
 
 /// Helper trait for the [alt()] combinator.
 ///
