+++ src/combinator/mod.rs
@@ -153,13 +153,13 @@ use crate::lib::std::boxed::Box;
 use crate::error::{ErrorKind, FromExternalError, ParseError};
 use crate::input::{AsChar, InputIter, InputLength, InputTakeAtPosition, ParseTo};
 use crate::input::{Compare, CompareResult, Offset, Slice};
-use crate::internal::*;
 use crate::lib::std::borrow::Borrow;
 use crate::lib::std::convert;
 #[cfg(feature = "std")]
 use crate::lib::std::fmt::Debug;
 use crate::lib::std::mem::transmute;
 use crate::lib::std::ops::{Range, RangeFrom, RangeTo};
+use crate::*;
 
 #[cfg(test)]
 mod tests;
