+++ src/error.rs
@@ -476,8 +476,8 @@
 //!
 //! You can go further with the [nom-trace crate](https://github.com/rust-bakery/nom-trace)
 
-use crate::internal::Parser;
 use crate::lib::std::fmt;
+use crate::Parser;
 
 /// This trait must be implemented by the error type of a nom parser.
 ///
@@ -693,7 +693,7 @@ impl<I: fmt::Display> fmt::Display for VerboseError<I> {
 #[cfg(feature = "std")]
 impl<I: fmt::Debug + fmt::Display> std::error::Error for VerboseError<I> {}
 
-use crate::internal::{Err, IResult};
+use crate::{Err, IResult};
 
 /// Create a new error from an input position, a static string and an existing error.
 /// This is used mainly in the [context] combinator, to add user friendly information
