+++ src/multi/mod.rs
@@ -6,9 +6,9 @@ mod tests;
 use crate::error::ErrorKind;
 use crate::error::ParseError;
 use crate::input::{InputLength, InputTake, ToUsize};
-use crate::internal::{Err, IResult, Needed, Parser};
 #[cfg(feature = "alloc")]
 use crate::lib::std::vec::Vec;
+use crate::{Err, IResult, Needed, Parser};
 use core::num::NonZeroUsize;
 
 /// Don't pre-allocate more than 64KiB when calling `Vec::with_capacity`.
